# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMA API",
        "description": "Sistem Informasi Manajemen Amanah — REST API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/reports/fund-balances": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Saldo seluruh Dana Amanah",
                "operationId": "3419387b1b61e9ffcff02f73c5418707",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/account-balances": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Saldo seluruh kas/bank",
                "operationId": "48941d9a2921539e0c50dd0a3435a64e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/ledger": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Buku besar (ledger)",
                "operationId": "ba0c690f5aa3af6f17086e3c0120312c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/reconciliation-summary": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Rekonsiliasi global Amanah",
                "operationId": "06ce0261046213b6c5972eaae4adb98f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reports/fund-statement": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Mutasi per Dana Amanah",
                "operationId": "8c3df663928f5825c9600f6de7a91b79",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "nullable": true
                    },
                    "meta": {
                        "nullable": true
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token dari POST /api/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}